requirements_eval.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
colorama==0.4.6
coloredlogs==15.0.1
DataProperty==1.0.1
datasets==2.13.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.7
einops==0.6.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
identify==2.5.25
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
jsonlines==3.1.0
jstyleson==0.0.2
lit==16.0.6
lm-eval==0.3.0
MarkupSafe==2.1.3
mbstrdecoder==1.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==2.8.8
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
onnx==1.14.0
openai==0.27.8
opencv-python==4.8.0.74
#openvino @ file:///home/aanuf/libs/wheels/openvino-2023.1.0.dev20230728-11840-cp38-cp38-manylinux_2_31_x86_64.whl#sha256=62ff3bb1387037bb3eaea257ed492f478933563fb184362b52d85b9ef5186215
#openvino-dev @ file:///home/aanuf/libs/wheels/openvino_dev-2023.1.0.dev20230728-11840-py3-none-any.whl#sha256=c1cadfd3b0465cb388e8cc52ed4f8cf000c0524e3f1c47d05b9fe7367b0a9778
openvino-telemetry==2023.1.0
optimum==1.9.1
optimum-intel==1.9.3
packaging==23.1
pandas==2.0.3
pathvalidate==3.1.0
peft==0.4.0
Pillow==10.0.0
pkg_resources==0.0.0
platformdirs==3.9.1
portalocker==2.7.0
pre-commit==3.3.3
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pybind11==2.11.1
pycountry==22.3.5
pytablewriter==1.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rouge-score==0.1.2
sacrebleu==1.5.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sqlitedict==2.1.0
sympy==1.12
tabledata==1.3.1
tcolorpy==0.1.3
texttable==1.6.7
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
tqdm-multiprocess==0.0.11
transformers==4.31.0
triton==2.0.0
typepy==1.3.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
